fix bag with merge
