+++ src/main/java/faang/school/analytics/config/redis/RedisConfiguration.java
@@ -1,5 +1,6 @@
 package faang.school.analytics.config.redis;
 
+import faang.school.analytics.listeners.CommentEventListener;
 import faang.school.analytics.listeners.RecommendationEventListener;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -7,9 +8,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
 @RequiredArgsConstructor
@@ -21,6 +24,8 @@ public class RedisConfiguration {
     private String host;
     @Value("${spring.data.redis.port}")
     private int port;
+    @Value("${spring.data.redis.channel.comment}")
+    private String commentChannelName;
 
 
     private final RecommendationEventListener recommendationEventListener;
@@ -32,13 +37,42 @@ public class RedisConfiguration {
     }
 
     @Bean
-    public RedisMessageListenerContainer redisContainer(JedisConnectionFactory connectionFactory) {
+    public RedisTemplate<String, Object> redisTemplate() {
+        RedisTemplate<String, Object> template = new RedisTemplate<>();
+        template.setConnectionFactory(connectionFactory());
+        template.setKeySerializer(new StringRedisSerializer());
+        template.setValueSerializer(new StringRedisSerializer());
+        return template;
+    }
+
+    @Bean
+    public MessageListenerAdapter commentListener(CommentEventListener commentEventListener) {
+        return new MessageListenerAdapter(commentEventListener);
+    }
+
+    @Bean
+    public ChannelTopic commentTopic() {
+        return new ChannelTopic(commentChannelName);
+    }
+
+
+    @Bean
+    public RedisMessageListenerContainer redisContainer(JedisConnectionFactory connectionFactory, MessageListenerAdapter commentListener) {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
         container.setConnectionFactory(connectionFactory);
 
         MessageListenerAdapter recommendationListenerAdapter = new MessageListenerAdapter(recommendationEventListener);
         container.addMessageListener(recommendationListenerAdapter, new ChannelTopic(userRecommendationChannel));
+        container.addMessageListener(commentListener, commentTopic());
 
         return container;
     }
+
+//    @Bean
+//    public RedisMessageListenerContainer redisContainer(MessageListenerAdapter commentListener) {
+//        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+//        container.setConnectionFactory(connectionFactory());
+//        container.addMessageListener(commentListener, commentTopic());
+//        return container;
+//    }
 }
