+++ src/main/java/faang/school/analytics/listeners/CommentEventListener.java
@@ -1,11 +1,11 @@
-package faang.school.analytics.listener;
+package faang.school.analytics.listeners;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.analytics.dto.AnalyticsEventMapper;
 import faang.school.analytics.dto.CommentEvent;
+import faang.school.analytics.mapper.AnalyticsMapper;
 import faang.school.analytics.model.AnalyticsEvent;
 import faang.school.analytics.model.EventType;
-import faang.school.analytics.service.AnalyticsEventService;
+import faang.school.analytics.service.AnalyticsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
@@ -19,17 +19,17 @@ import java.io.IOException;
 @RequiredArgsConstructor
 public class CommentEventListener implements MessageListener {
 
-    private final AnalyticsEventMapper analyticsEventMapper;
+    private final AnalyticsMapper analyticsMapper;
 
     private final ObjectMapper objectMapper;
-    private final AnalyticsEventService analyticsEventService;
+    private final AnalyticsService analyticsEventService;
 
     @Override
     public void onMessage(Message message, byte[] pattern) {
         CommentEvent event = null;
         try {
             event = objectMapper.readValue(message.getBody(), CommentEvent.class);
-            AnalyticsEvent mapperAnalyticsEvent = analyticsEventMapper.toAnalyticsEvent(event);
+            AnalyticsEvent mapperAnalyticsEvent = analyticsMapper.toAnalyticsEvent(event);
             mapperAnalyticsEvent.setEventType(EventType.POST_COMMENT);
             AnalyticsEvent analyticsEvent = analyticsEventService.addAnalyticsEvent(mapperAnalyticsEvent);
             log.info("CommentEventListener CommentEvent: {}", event);
