+++ src/main/java/faang/school/analytics/mapper/AnalyticsMapper.java
@@ -2,8 +2,10 @@ package faang.school.analytics.mapper;
 
 import faang.school.analytics.dto.AnalyticsCreateEventDto;
 import faang.school.analytics.dto.AnalyticsEventDto;
+import faang.school.analytics.dto.CommentEvent;
 import faang.school.analytics.model.AnalyticsEvent;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
@@ -12,4 +14,9 @@ public interface AnalyticsMapper {
     AnalyticsEvent toEntity(AnalyticsEventDto analyticsEventDto);
     AnalyticsEvent toEntity(AnalyticsCreateEventDto analyticsCreateEventDto);
     AnalyticsEventDto toDto(AnalyticsEvent analyticsEvent);
+
+    @Mapping(source = "authorCommentId", target = "actorId")
+    @Mapping(source = "authorPostId", target = "receiverId")
+    @Mapping(source = "createdAt", target = "receivedAt")
+    AnalyticsEvent toAnalyticsEvent(CommentEvent commentEvent);
 }
